Moves quote fetching to async thunk
