+++ packages/synapse-interface/slices/bridgeQuote/reducer.ts
@@ -2,15 +2,18 @@ import { createSlice, PayloadAction } from '@reduxjs/toolkit'
 
 import { EMPTY_BRIDGE_QUOTE } from '@/constants/bridge'
 import { type BridgeQuote } from '@/utils/types'
+import { fetchBridgeQuote } from './thunks'
 
 export interface BridgeQuoteState {
   bridgeQuote: BridgeQuote
   isLoading: boolean
+  error: any
 }
 
 export const initialState: BridgeQuoteState = {
   bridgeQuote: EMPTY_BRIDGE_QUOTE,
   isLoading: false,
+  error: null,
 }
 
 export const bridgeQuoteSlice = createSlice({
@@ -20,16 +23,33 @@ export const bridgeQuoteSlice = createSlice({
     setIsLoading: (state, action: PayloadAction<boolean>) => {
       state.isLoading = action.payload
     },
-    setBridgeQuote: (state, action: PayloadAction<BridgeQuote>) => {
-      state.bridgeQuote = action.payload
-    },
     resetBridgeQuote: (state) => {
       state.bridgeQuote = initialState.bridgeQuote
     },
   },
+  extraReducers: (builder) => {
+    builder
+      .addCase(fetchBridgeQuote.pending, (state) => {
+        // state.status = FetchState.LOADING
+        state.isLoading = true
+      })
+      .addCase(
+        fetchBridgeQuote.fulfilled,
+        (state, action: PayloadAction<BridgeQuote>) => {
+          state.bridgeQuote = action.payload
+          // state.status = FetchState.VALID
+          state.isLoading = false
+        }
+      )
+      .addCase(fetchBridgeQuote.rejected, (state, action) => {
+        // state.error = action.payload
+        state.bridgeQuote = EMPTY_BRIDGE_QUOTE
+        // state.status = FetchState.INVALID
+        state.isLoading = false
+      })
+  },
 })
 
-export const { setBridgeQuote, resetBridgeQuote, setIsLoading } =
-  bridgeQuoteSlice.actions
+export const { resetBridgeQuote, setIsLoading } = bridgeQuoteSlice.actions
 
 export default bridgeQuoteSlice.reducer
